COM-2727 - create eventReducer file
